refactor(ItemsPage): Make BestItemsSection and AllItemsSection to clarify each one's operation
