+++ src/components/AllItemsSection.js
@@ -1,0 +1,76 @@
+import { useEffect, useState } from "react";
+import { MOBILE, NUM_ALL_ITEMS, RECENT } from "../utils/magicLiterals";
+import { getItems } from "../api";
+import ItemListNav from "./ItemListNav.js";
+import ItemListNavMobile from "./ItemListNavMobile.js";
+import ItemList from "./ItemList";
+import Pagination from "./Pagination.js";
+
+function AllItemsSection({ device }) {
+  const [allItems, setAllItems] = useState([]);
+  const [pageInfo, setPageInfo] = useState({
+    currentPage: 1,
+    totalCount: 0,
+    totalPage: 1,
+  });
+  const [order, setOrder] = useState(RECENT);
+
+  const handleInitialLoad = async () => {
+    const totalCount = Number(await handleLoad());
+    const totalPage = Math.ceil(totalCount / NUM_ALL_ITEMS[device]);
+
+    setPageInfo((prevPageInfo) => ({
+      ...prevPageInfo,
+      totalCount,
+      totalPage,
+    }));
+  };
+
+  const handleLoad = async () => {
+    const { list: newAllItems, totalCount: totalCountString } = await getItems({
+      page: pageInfo.currentPage,
+      pageSize: NUM_ALL_ITEMS[device],
+      orderBy: order,
+    });
+    setAllItems(newAllItems);
+    return totalCountString;
+  };
+
+  useEffect(() => {
+    handleInitialLoad();
+  }, []);
+
+  useEffect(() => {
+    const totalPage = Math.ceil(pageInfo.totalCount / NUM_ALL_ITEMS[device]);
+    setPageInfo((prevPageInfo) => ({
+      ...prevPageInfo,
+      currentPage: 1, // device 바뀌었으니 다시 첫페이지로
+      totalPage,
+    }));
+  }, [device]);
+
+  useEffect(() => {
+    setPageInfo((prevPageInfo) => ({
+      ...prevPageInfo,
+      currentPage: 1, // order 바뀌었으니 다시 첫페이지로
+    }));
+  }, [order]);
+
+  useEffect(() => {
+    handleLoad();
+  }, [pageInfo]);
+
+  return (
+    <section id="all-items">
+      {device === MOBILE ? (
+        <ItemListNavMobile order={order} setOrder={setOrder} />
+      ) : (
+        <ItemListNav order={order} setOrder={setOrder} />
+      )}
+      <ItemList items={allItems} />
+      <Pagination pageInfo={pageInfo} setPageInfo={setPageInfo} />
+    </section>
+  );
+}
+
+export default AllItemsSection;
